Make the `ActionSelector` public.
