+++ src/main/java/io/jenkins/plugins/echarts/GenericBuildActionIterator.java
@@ -62,58 +62,6 @@ public class GenericBuildActionIterator<A extends BuildAction<?>, R> implements 
     }
 
     /**
-     * Selects a specific action from the all actions that are attached to the given build. The action is selected by a
-     * generic predicate that works on the expected concrete action type. If the baseline build does not contain the
-     * action then previous builds will be inspected until the action is found.
-     *
-     * @param <T>
-     *         the type of the action to select
-     */
-    static class ActionSelector<T extends BuildAction<?>> implements Function<Run<?, ?>, Optional<T>> {
-        private final Class<T> actionType;
-        private final Predicate<? super T> predicate;
-
-        /**
-         * Creates a new instance of {@link ActionSelector}. This selector will select the first action of the given
-         * type that matches.
-         *
-         * @param actionType
-         *         the type of the action to select
-         */
-        ActionSelector(final Class<T> actionType) {
-            this(actionType, action -> true);
-        }
-
-        /**
-         * Creates a new instance of {@link ActionSelector}.
-         *
-         * @param actionType
-         *         the type of the action to select
-         * @param predicate
-         *         the predicate that selects the action (if there are multiple actions of the same type)
-         */
-        ActionSelector(final Class<T> actionType, final Predicate<? super T> predicate) {
-            this.actionType = actionType;
-            this.predicate = predicate;
-        }
-
-        @Override
-        public Optional<T> apply(final Run<?, ?> baseline) {
-            for (Run<?, ?> run = baseline; run != null; run = run.getPreviousBuild()) {
-                Optional<T> action = run.getActions(actionType)
-                        .stream()
-                        .filter(predicate)
-                        .findAny();
-                if (action.isPresent()) {
-                    return action;
-                }
-            }
-
-            return Optional.empty();
-        }
-    }
-
-    /**
      * An iterable that provides an iterator for specific values of build that should be rendered in a trend chart. The
      * build values must be stored in a concrete subclass of {@link BuildAction}. The property that contains the build value
      * is selected by a generic predicate.
